requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
tk>=0.1.0
matplotlib>=3.5.0
pycocotools>=2.0.4 